# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bankit fintech API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions (filtered by role)
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [verified, pending, failed]
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Transactions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}/receipt:
    get:
      operationId: getTransactionReceipt
      tags: [transactions]
      summary: Generate/get receipt for a transaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receipt data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
  /transactions/verify:
    post:
      operationId: verifyPayment
      tags: [transactions]
      summary: Verify a payment (sales rep / account officer)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentVerifyInput"
      responses:
        "200":
          description: Verification result (role-masked)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResult"
  /transactions/recent:
    get:
      operationId: getRecentTransactions
      tags: [transactions]
      summary: Get recent transactions (last 10, role-filtered)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recent transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
  /transactions/summary:
    get:
      operationId: getTransactionSummary
      tags: [transactions]
      summary: Get transaction summary stats (role-gated)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionSummary"

  # ── Public verification endpoint ───────────────────────────
  /verify/{cplid}:
    get:
      operationId: publicVerifyCplid
      tags: [verify]
      summary: Public payment verification by CPLID
      parameters:
        - name: cplid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicVerifyResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Transfers ──────────────────────────────────────────────
...
  /reports/revenue:
    get:
      operationId: getRevenueReport
      tags: [reports]
      summary: Get revenue report (account officer + admin)
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly, annual]
      responses:
        "200":
          description: Revenue report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueReport"
  /reports/dashboard:
    get:
      operationId: getDashboardStats
      tags: [reports]
      summary: Get dashboard stats (role-gated fields)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /reports/export:
    get:
      operationId: exportReport
      tags: [reports]
      summary: Export report as JSON (account officer + admin)
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [revenue, transactions, audit]
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly, annual]
      responses:
        "200":
          description: Report export data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportData"

  # ── Tax ────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Transaction:
      type: object
      required: [id, cplid, amount, customerName, customerAccountNumber, bankName, date, status, referenceNumber]
      properties:
        id:
          type: integer
        cplid:
          type: string
          description: "BKT-XXXXXXXX unique identifier"
        amount:
          type: number
        customerName:
          type: string
        customerAccountNumber:
          type: string
        bankName:
          type: string
        date:
          type: string
        status:
          type: string
          enum: [verified, pending, failed]
        referenceNumber:
          type: string
        description:
          type: ["string", "null"]
        verifiedBy:
          type: ["integer", "null"]
        verifiedAt:
          type: ["string", "null"]
    TransactionList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    TransactionSummary:
      type: object
      required: [verified, pending, failed, totalCount]
      properties:
        verified:
          type: integer
        pending:
          type: integer
        failed:
          type: integer
        totalCount:
          type: integer
        todayCount:
          type: integer
    PaymentVerifyInput:
      type: object
      required: [query]
      properties:
        query:
          type: string
          description: CPLID, invoice number, reference number, or phone number
    VerificationResult:
      type: object
      required: [found, status]
      properties:
        found:
          type: boolean
        status:
          type: ["string", "null"]
          enum: [verified, pending, failed, null]
        customerName:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        paymentDate:
          type: ["string", "null"]
        referenceNumber:
          type: ["string", "null"]
        cplid:
          type: ["string", "null"]
    PublicVerifyResult:
      type: object
      required: [status, customerName, amount, paymentDate, companyName]
      properties:
        status:
          type: string
          enum: [verified, pending, failed]
        customerName:
          type: string
        amount:
          type: number
        paymentDate:
          type: string
        companyName:
          type: string
        cplid:
          type: string
    Receipt:
      type: object
      required: [id, transactionId, cplid, customerName, amount, status, issuedAt]
      properties:
        id:
          type: integer
        transactionId:
          type: integer
        cplid:
          type: string
        customerName:
          type: string
        amount:
          type: number
        status:
          type: string
        referenceNumber:
          type: string
        issuedAt:
          type: string
        qrCodeData:
          type: ["string", "null"]

    # ── Transfers ────────────────────────────────────────
...
    RevenueReport:
      type: object
      required: [period, totalRevenue, periodBreakdown]
      properties:
        period:
          type: string
        totalRevenue:
          type: number
        transactionCount:
          type: integer
        periodBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/RevenueDataPoint"
    RevenueDataPoint:
      type: object
      required: [label, amount, count]
      properties:
        label:
          type: string
        amount:
          type: number
        count:
          type: integer
    DashboardStats:
      type: object
      required: [role]
      properties:
        role:
          type: string
        recentVerifications:
          type: integer
        pendingVerifications:
          type: integer
        todayRevenue:
          type: ["number", "null"]
        weeklyRevenue:
          type: ["number", "null"]
        monthlyRevenue:
          type: ["number", "null"]
        grossRevenue:
          type: ["number", "null"]
        annualRevenue:
          type: ["number", "null"]
        totalUsers:
          type: ["integer", "null"]
        pendingTransfers:
          type: ["integer", "null"]
        totalTransactions:
          type: integer
        verifiedToday:
          type: integer
    ExportData:
      type: object
      required: [type, period, generatedAt, data]
      properties:
        type:
          type: string
        period:
          type: string
        generatedAt:
          type: string
        data:
          type: array
          items:
            type: object

    # ── Tax ──────────────────────────────────────────────
...